Fix parent width calculation in useContainerMQ()
width() returns the wrong value because it is called before the
necessary processes have finished.

Call of setTimeout() without timeout adds the call of its callback
function to the end of the callback queue.

Thus setTimeout() defers the call of width() until all necessary
processes finish so width() returns the correct value.